@@ -40,60 +41,61 @@
         ]
     },
     "devDependencies": {
-        "@babel/cli": "^7.6.2",
-        "@babel/core": "^7.6.2",
-        "@babel/node": "^7.6.2",
-        "@babel/polyfill": "^7.6.0",
-        "@babel/preset-env": "^7.6.2",
-        "@babel/preset-react": "^7.0.0",
-        "babel-loader": "^8.0.6",
-        "concurrently": "^5.0.2",
+        "@babel/cli": "^7.10.5",
+        "@babel/core": "^7.10.5",
+        "@babel/node": "^7.10.5",
+        "@babel/polyfill": "^7.10.4",
+        "@babel/preset-env": "^7.10.4",
+        "@babel/preset-react": "^7.10.4",
+        "babel-loader": "^8.1.0",
+        "concurrently": "^5.2.0",
         "cross-env": "^6.0.3",
-        "css-loader": "^3.2.0",
-        "electron": "^6.0.11",
-        "electron-packager": "^14.1.1",
+        "css-loader": "^3.6.0",
+        "electron": "^9.1.0",
+        "electron-packager": "^15.0.0",
         "express": "^4.16.2",
-        "style-loader": "^1.0.0",
-        "webpack": "^4.41.0",
-        "webpack-cli": "^3.3.9",
+        "style-loader": "^1.2.1",
+        "webpack": "^4.44.1",
+        "webpack-cli": "^3.3.12",
         "webpack-dev-middleware": "^3.7.2",
         "webpack-hot-middleware": "^2.25.0"
     },
     "dependencies": {
-        "@fortawesome/fontawesome-free": "^5.1.1",
-        "async": "^2.6.0",
-        "bootstrap": "^3.3.7",
-        "clsx": "^1.0.4",
-        "dagre": "^0.7.4",
+        "@fortawesome/fontawesome-free": "^5.14.0",
+        "async": "^2.6.3",
+        "bootstrap": "^3.4.1",
+        "clsx": "^1.1.1",
+        "dagre": "^0.8.5",
         "electron-store": "^1.3.0",
-        "eventemitter2": "^4.1.0",
+        "eventemitter2": "^6.4.3",
         "fontfaceobserver": "^2.0.13",
-        "framer-motion": "^1.8.3",
-        "handlebars": "^4.4.2",
+        "framer-motion": "^1.11.1",
+        "handlebars": "^4.7.6",
         "he": "^1.2.0",
-        "image-size": "^0.6.3",
-        "image-type": "^3.0.0",
+        "image-size": "^0.8.3",
+        "image-type": "^4.1.0",
         "is-zip-file": "^1.0.2",
-        "jquery": "^3.2.1",
+        "jquery": "^3.5.1",
         "linkurious": "^1.5.1",
         "md5-file": "^4.0.0",
         "mustache": "^2.3.0",
-        "neo4j-driver": "^1.7.5",
+        "neo4j-driver": "^4.1.0",
         "prop-types": "^15.7.2",
-        "react": "^16.8.6",
+        "react": "^16.13.1",
         "react-alert": "^5.5.0",
         "react-alert-template-basic": "^1.0.0",
-        "react-bootstrap": "^0.32.4",
-        "react-bootstrap-typeahead": "^3.4.7",
-        "react-dom": "^16.8.6",
-        "react-draggable": "^4.2.0",
+        "react-bootstrap": "^0.33.1",
+        "react-bootstrap-typeahead": "^4.2.3",
+        "react-dom": "^16.13.1",
+        "react-draggable": "^4.4.3",
         "react-if": "^3.4.3",
-        "react-images": "^1.0.0",
+        "react-images": "^1.1.7",
         "react-photo-gallery": "^8.0.0",
-        "react-transition-group": "^4.2.0",
+        "react-transition-group": "^4.4.1",
         "sanitize-filename": "^1.6.3",
-        "stream-json": "^1.1.4",
-        "unzipper": "^0.8.9",
-        "uuid": "^3.3.2"
+        "stream-chain": "^2.2.3",
+        "stream-json": "^1.7.1",
+        "unzipper": "^0.10.11",
+        "uuid": "^3.4.0"
     }
 }
