@@ -1,55 +1,59 @@
-import React, { Component } from 'react';
-import PrebuiltQueryNode from './PrebuiltQueryNode';
-import { If, Then, Else } from 'react-if';
+import React, { useEffect, useState } from 'react';
+import PropTypes from 'prop-types';
 import { remote } from 'electron';
 const { app } = remote;
-import { join } from 'path';
+import path from 'path';
+import fs from 'fs';
 import { platform } from 'process';
 import { exec } from 'child_process';
+import PrebuiltQueryNode from './PrebuiltQueryNode';
+import styles from './PrebuiltQueries.module.css';
+import { useContext } from 'react';
+import { AppContext } from '../../../AppContext';
 
-export default class PrebuiltQueriesDisplay extends Component {
-    constructor() {
-        super();
+const PrebuiltQueriesDisplay = () => {
+    const [queries, setQueries] = useState([]);
+    const [custom, setCustom] = useState([]);
+    const context = useContext(AppContext);
 
-        this.state = {
-            queries: [],
-            custom: [],
-        };
-    }
+    useEffect(() => {
+        readCustom();
+        readBase();
+    }, []);
 
-    componentWillMount() {
-        $.ajax({
-            url: join(app.getPath('userData'), '/customqueries.json'),
-            type: 'GET',
-            success: function(response) {
-                var x = JSON.parse(response);
-                var y = [];
+    const readCustom = async () => {
+        let filePath = path.join(
+            app.getPath('userData'),
+            '/customqueries.json'
+        );
+        fs.readFile(filePath, 'utf8', (err, data) => {
+            let j = JSON.parse(data);
+            var y = [];
+            j.queries.forEach((query) => {
+                y.push(query);
+            });
 
-                $.each(x.queries, function(_, el) {
-                    y.push(el);
-                });
-
-                this.setState({ custom: y });
-            }.bind(this),
+            setCustom(y);
         });
+    };
 
+    const readBase = async () => {
         $.ajax({
             url: 'src/components/SearchContainer/Tabs/PrebuiltQueries.json',
             type: 'GET',
-            success: function(response) {
-                var x = JSON.parse(response);
+            success: function (response) {
                 var y = [];
 
-                $.each(x.queries, function(_, el) {
+                $.each(response.queries, function (_, el) {
                     y.push(el);
                 });
 
-                this.setState({ queries: y });
-            }.bind(this),
+                setQueries(y);
+            },
         });
-    }
+    };
 
-    getCommandLine() {
+    const getCommandLine = () => {
         switch (platform) {
             case 'darwin':
                 return 'open';
@@ -60,81 +64,58 @@ export default class PrebuiltQueriesDisplay extends Component {
             default:
                 return 'xdg-open';
         }
-    }
+    };
 
-    editCustom() {
+    const editCustom = () => {
         exec(
-            this.getCommandLine() +
+            getCommandLine() +
                 ' "' +
                 join(app.getPath('userData'), '/customqueries.json') +
                 '"'
         );
-    }
+    };
 
-    refreshCustom() {
-        $.ajax({
-            url: join(app.getPath('userData'), '/customqueries.json'),
-            type: 'GET',
-            success: function(response) {
-                var x = JSON.parse(response);
-                var y = [];
+    const refreshCustom = () => {
+        readCustom();
+    };
 
-                $.each(x.queries, function(index, el) {
-                    y.push(el);
-                });
-
-                this.setState({ custom: y });
-            }.bind(this),
-        });
-    }
-
-    render() {
-        return (
-            <div>
-                <h3>Pre-Built Analytics Queries</h3>
-                <div className='query-box'>
-                    {this.state.queries.map(function(a) {
-                        return <PrebuiltQueryNode key={a.name} info={a} />;
-                    })}
-                </div>
-                <h3>
-                    Custom Queries
-                    <i
-                        className='glyphicon glyphicon-pencil customQueryGlyph'
-                        data-toggle='tooltip'
-                        title='Edit Queries'
-                        onClick={this.editCustom.bind(this)}
-                    />
-                    <i
-                        className='glyphicon glyphicon-refresh customQueryGlyph'
-                        onClick={this.refreshCustom.bind(this)}
-                        style={{ paddingLeft: '5px' }}
-                        data-toggle='tooltip'
-                        title='Refresh Queries'
-                    />
-                </h3>
-                <div className='query-box'>
-                    <If condition={this.state.custom.length === 0}>
-                        <Then>
-                            <div>No user defined queries.</div>
-                        </Then>
-                        <Else>
-                            {() => (
-                                <div>
-                                    {this.state.custom.map(function(a) {
-                                        return (
-                                            <PrebuiltQueryNode
-                                                key={a.name}
-                                                info={a}
-                                            />
-                                        );
-                                    })}
-                                </div>
-                            )}
-                        </Else>
-                    </If>
-                </div>
+    return (
+        <div className={context.darkMode ? styles.dark : null}>
+            <h3>Pre-Built Analytics Queries</h3>
+            <div className='query-box'>
+                {queries.map(function (a) {
+                    return <PrebuiltQueryNode key={a.name} info={a} />;
+                })}
             </div>
-        );
-    }
-}
+            <h3>
+                Custom Queries
+                <i
+                    className='glyphicon glyphicon-pencil customQueryGlyph'
+                    data-toggle='tooltip'
+                    title='Edit Queries'
+                    onClick={editCustom}
+                />
+                <i
+                    className='glyphicon glyphicon-refresh customQueryGlyph'
+                    onClick={refreshCustom}
+                    style={{ paddingLeft: '5px' }}
+                    data-toggle='tooltip'
+                    title='Refresh Queries'
+                />
+            </h3>
+            <div className='query-box'>
+                {custom.length === 0 && <div>No user defined queries.</div>}
+                {custom.length > 0 && (
+                    <div>
+                        {custom.map(function (a) {
+                            return <PrebuiltQueryNode key={a.name} info={a} />;
+                        })}
+                    </div>
+                )}
+            </div>
+        </div>
+    );
+};
+
+PrebuiltQueriesDisplay.propTypes = {};
+export default PrebuiltQueriesDisplay;
