@@ -1,110 +1,3 @@
-// import React, { Component } from 'react';
-
-// export default class ExportContainer extends Component {
-//     constructor() {
-//         super();
-
-//         this.state = {
-//             jsonActive: true,
-//             imageActive: false,
-//         };
-//     }
-
-//     componentDidMount() {
-//         $(this.refs.outer).fadeToggle(0);
-//         $(this.refs.outer).draggable();
-//         emitter.on('showExport', this._show.bind(this));
-//     }
-
-//     _dismiss() {
-//         $(this.refs.outer).fadeToggle(false);
-//     }
-
-//     _show() {
-//         closeTooltip();
-//         $(this.refs.outer).fadeToggle(true);
-//     }
-
-//     _jsonClick() {
-//         this.setState({ jsonActive: true, imageActive: false });
-//     }
-
-//     _imageClick() {
-//         this.setState({ jsonActive: false, imageActive: true });
-//     }
-
-//     _buttonClick() {
-//         emitter.emit('export', this.state.jsonActive ? 'json' : 'image');
-//     }
-
-//     render() {
-//         return (
-//             <div className='floating-div' ref='outer'>
-//                 <div className='panel-heading'>
-//                     Export Graph
-//                     <button
-//                         type='button'
-//                         className='close'
-//                         onClick={this._dismiss.bind(this)}
-//                         aria-label='Close'
-//                     >
-//                         <span aria-hidden='true'>&times;</span>
-//                     </button>
-//                 </div>
-
-//                 <div className='panel-body'>
-//                     <div className='list-group'>
-//                         <a
-//                             href='#'
-//                             onClick={this._jsonClick.bind(this)}
-//                             className={
-//                                 this.state.jsonActive
-//                                     ? 'list-group-item active'
-//                                     : 'list-group-item'
-//                             }
-//                         >
-//                             <h4 className='list-group-item-heading'>
-//                                 Export to JSON
-//                             </h4>
-//                             <p className='list-group-item-text'>
-//                                 Use this format to export data and re-import it
-//                                 later
-//                             </p>
-//                         </a>
-
-//                         <a
-//                             href='#'
-//                             onClick={this._imageClick.bind(this)}
-//                             className={
-//                                 this.state.imageActive
-//                                     ? 'list-group-item active'
-//                                     : 'list-group-item'
-//                             }
-//                         >
-//                             <h4 className='list-group-item-heading'>
-//                                 Export to Image
-//                             </h4>
-//                             <p className='list-group-item-text'>
-//                                 Use this format to export data and view it as an
-//                                 image
-//                             </p>
-//                         </a>
-//                     </div>
-
-//                     <div style={{ textAlign: 'center' }}>
-//                         <button
-//                             onClick={this._buttonClick.bind(this)}
-//                             className='btn btn-lg'
-//                         >
-//                             Export Data
-//                         </button>
-//                     </div>
-//                 </div>
-//             </div>
-//         );
-//     }
-// }
-
 import React, { useEffect, useState, useContext } from 'react';
 import PropTypes from 'prop-types';
 import PoseContainer from '../PoseContainer';
@@ -113,11 +6,14 @@ import styles from './ExportContainer.module.css';
 import { Panel, Button, ListGroup, ListGroupItem } from 'react-bootstrap';
 import { AppContext } from '../../AppContext';
 import clsx from 'clsx';
+import { useDragControls } from 'framer-motion';
+import { Table } from 'react-bootstrap';
 
 const ExportContainer = () => {
     const [jsonActive, setJsonActive] = useState(true);
     const [imageActive, setImageActive] = useState(false);
     const [visible, setVisible] = useState(false);
+    const dragControl = useDragControls();
 
     const context = useContext(AppContext);
 
@@ -130,18 +26,12 @@ const ExportContainer = () => {
         setVisible(false);
     };
 
-    const jsonClicked = () => {
-        setJsonActive(true);
-        setImageActive(false);
+    const startPNGExport = () => {
+        emitter.emit('export', 'image');
     };
 
-    const imageClicked = () => {
-        setJsonActive(false);
-        setImageActive(true);
-    };
-
-    const startExport = () => {
-        emitter.emit('export', jsonActive ? 'json' : 'image');
+    const startJSONExport = () => {
+        emitter.emit('export', 'json');
     };
 
     useEffect(() => {
@@ -152,54 +42,57 @@ const ExportContainer = () => {
     }, []);
 
     return (
-        <Draggable handle={'.panel-heading'}>
-            <PoseContainer
-                visible={visible}
-                className={clsx(
-                    styles.container,
-                    context.darkMode ? styles.dark : null
-                )}
-            >
-                <Panel>
-                    <Panel.Heading>
-                        Export Graph
-                        <Button
-                            onClick={handleClose}
-                            className='close'
-                            aria-label='Close'
-                        >
-                            <span aria-hidden='true'>&times;</span>
-                        </Button>
-                    </Panel.Heading>
+        <PoseContainer
+            visible={visible}
+            className={clsx(
+                styles.container,
+                context.darkMode ? styles.dark : null
+            )}
+            dragHandle={dragControl}
+        >
+            <Panel>
+                <Panel.Heading onMouseDown={(e) => dragControl.start(e)}>
+                    Export Graph
+                    <Button
+                        onClick={handleClose}
+                        className='close'
+                        aria-label='Close'
+                    >
+                        <span aria-hidden='true'>x</span>
+                    </Button>
+                </Panel.Heading>
 
-                    <Panel.Body>
-                        <ListGroup>
-                            <ListGroupItem
-                                header={'Export to JSON'}
-                                onClick={jsonClicked}
-                                active={jsonActive}
-                            >
-                                Use this format to export data and re-import it
-                                later
-                            </ListGroupItem>
-                            <ListGroupItem
-                                header={'Export to Image'}
-                                onClick={imageClicked}
-                                active={imageActive}
-                            >
-                                Use this format to export data and view it as an
-                                image
-                            </ListGroupItem>
-                        </ListGroup>
-                        <div className={styles.centerButton}>
-                            <Button onClick={startExport} size={'lg'}>
-                                Export Data
-                            </Button>
-                        </div>
-                    </Panel.Body>
-                </Panel>
-            </PoseContainer>
-        </Draggable>
+                <Panel.Body>
+                    <div
+                        className={
+                            context.darkMode ? styles.itemlistdark : null
+                        }
+                    >
+                        <Table>
+                            <thead></thead>
+                            <tbody className='searchable'>
+                                <tr
+                                    style={{ cursor: 'pointer' }}
+                                    onClick={startJSONExport}
+                                >
+                                    <td width='200px' align='center'>
+                                        Export to JSON
+                                    </td>
+                                </tr>
+                                <tr
+                                    style={{ cursor: 'pointer' }}
+                                    onClick={startPNGExport}
+                                >
+                                    <td width='200px' align='center'>
+                                        Export to PNG
+                                    </td>
+                                </tr>
+                            </tbody>
+                        </Table>
+                    </div>
+                </Panel.Body>
+            </Panel>
+        </PoseContainer>
     );
 };
 
