@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import { Button } from 'react-bootstrap';
 import clsx from 'clsx';
 import { motion } from 'framer-motion';
+import neo4j from 'neo4j-driver';
 
 const Login = () => {
     const [url, setUrl] = useState('bolt://localhost:7687');
@@ -111,7 +112,7 @@ const Login = () => {
 
         session
             .run('MATCH (n) RETURN n LIMIT 1')
-            .then(_ => {
+            .then((_) => {
                 setLoginRunning(false);
                 setLoginSuccess(true);
 
@@ -142,6 +143,18 @@ const Login = () => {
                     }
                 );
 
+                session = global.driver.session();
+                session
+                    .run(
+                        'CALL dbms.components() YIELD versions RETURN versions[0] AS version'
+                    )
+                    .then((result) => {
+                        let record = result.records[0];
+                        let version = record.get('version');
+                        global.neoVersion = version;
+                        session.close();
+                    });
+
                 setTimeout(() => {
                     setVisible(false);
                     setTimeout(() => {
@@ -149,7 +162,7 @@ const Login = () => {
                     }, 400);
                 }, 1500);
             })
-            .catch(error => {
+            .catch((error) => {
                 console.log(error);
                 if (error.message.includes('authentication failure')) {
                     setLoginEnabled(true);
@@ -233,7 +246,7 @@ const Login = () => {
 
         session
             .run('MATCH (n) RETURN n LIMIT 1')
-            .then(result => {
+            .then((result) => {
                 icon.removeClass();
                 icon.addClass(
                     'fa fa-check-circle green-icon-color form-control-feedback'
@@ -241,7 +254,7 @@ const Login = () => {
                 setLoginEnabled(true);
                 setUrl(tempUrl);
             })
-            .catch(error => {
+            .catch((error) => {
                 if (error.message.includes('WebSocket connection failure')) {
                     icon.removeClass();
                     icon.addClass(
@@ -267,7 +280,7 @@ const Login = () => {
             });
     };
 
-    const triggerLogin = event => {
+    const triggerLogin = (event) => {
         let key = event.keyCode ? event.KeyCode : event.which;
 
         if (key === 13) {
@@ -282,28 +295,26 @@ const Login = () => {
                 animate={visible ? 'visible' : 'hidden'}
                 initial={'hidden'}
             >
-                <img src='src/img/logo-white-transparent-full.png' />
+                <img src='src/img/logo-white-on-transparent.png' />
                 <div className='text-center'>
                     <span>Log in to Neo4j Database</span>
                 </div>
+                <br></br>
                 <form>
                     <div className='form-group has-feedback'>
                         <div className='input-group'>
-                            <span className='input-group-addon' id='dburladdon'>
-                                Database URL
-                            </span>
                             <input
-                                onFocus={function() {
+                                onFocus={function () {
                                     icon.tooltip('hide');
                                 }}
                                 onBlur={checkDatabaseExists}
-                                onChange={event => {
+                                onChange={(event) => {
                                     setUrl(event.target.value);
                                 }}
                                 type='text'
-                                className='form-control'
+                                className='form-control login-text'
                                 value={url}
-                                placeholder='bolt://localhost:7687'
+                                placeholder='Neo4j URL'
                                 aria-describedby='dburladdon'
                             />
                             <i
@@ -312,38 +323,29 @@ const Login = () => {
                             />
                         </div>
                         <div className='input-group spacing'>
-                            <span
-                                className='input-group-addon'
-                                id='dbuseraddon'
-                            >
-                                DB Username
-                            </span>
                             <input
                                 type='text'
                                 value={user}
                                 onKeyDown={triggerLogin}
-                                onChange={event => {
+                                onChange={(event) => {
                                     setUser(event.target.value);
                                 }}
-                                className='form-control'
-                                placeholder='neo4j'
+                                className='form-control login-text'
+                                placeholder='Neo4j Username'
                                 aria-describedby='dbuseraddon'
                             />
                         </div>
                         <div className='input-group spacing'>
-                            <span className='input-group-addon' id='dbpwaddon'>
-                                DB Password
-                            </span>
                             <input
                                 ref={passwordRef}
                                 value={password}
                                 onKeyDown={triggerLogin}
-                                onChange={event => {
+                                onChange={(event) => {
                                     setPassword(event.target.value);
                                 }}
                                 type='password'
-                                className='form-control'
-                                placeholder='neo4j'
+                                className='form-control login-text'
+                                placeholder='Neo4j Password'
                                 aria-describedby='dbpwaddon'
                             />
                         </div>
@@ -352,12 +354,12 @@ const Login = () => {
                                 <label>
                                     <input
                                         checked={save}
-                                        onChange={event =>
+                                        onChange={(event) =>
                                             setSave(event.target.checked)
                                         }
                                         type='checkbox'
                                     />
-                                    Save Password
+                                    <font color='white'>Save Password</font>
                                 </label>
                             </div>
                             <div className='buttoncontainer'>
