@@ -1,4 +1,4 @@
-import React, { useEffect, useState, useContext } from 'react';
+import React, { useRef, useEffect, useState, useContext } from 'react';
 import PropTypes from 'prop-types';
 import { withAlert } from 'react-alert';
 import NodeEditorRow from './NodeEditorRow.jsx';
@@ -8,6 +8,7 @@ import PoseContainer from '../PoseContainer';
 import Draggable from 'react-draggable';
 import clsx from 'clsx';
 import { AppContext } from '../../AppContext.jsx';
+import { useDragControls } from 'framer-motion';
 
 const NodeEditor = () => {
     const [name, setName] = useState('');
@@ -19,8 +20,9 @@ const NodeEditor = () => {
     const [newAttrType, setNewAttrType] = useState('boolean');
     const [hasError, setHasError] = useState(false);
     const context = useContext(AppContext);
+    const dragControl = useDragControls();
 
-    const getNodeData = node => {
+    const getNodeData = (node) => {
         setName(node.label);
         setType(node.type);
         setId(node.objectid);
@@ -29,11 +31,11 @@ const NodeEditor = () => {
         let session = driver.session();
         let statement = `MATCH (n:${node.type} {objectid: $id}) RETURN n`;
 
-        session.run(statement, { id: node.objectid }).then(result => {
+        session.run(statement, { id: node.objectid }).then((result) => {
             let props = result.records[0]._fields[0].properties;
             let label = props.name;
             delete props.name;
-            setName(label);
+            setName(label || props.objectid);
             setProperties(props);
             session.close();
         });
@@ -58,11 +60,11 @@ const NodeEditor = () => {
 
         let session = driver.session();
         let statement = `MATCH (n:${type} {objectid:  $id}) SET n.${newAttrName}=$newprop RETURN n`;
-        session.run(statement, { id: id, newprop: newValue }).then(result => {
+        session.run(statement, { id: id, newprop: newValue }).then((result) => {
             let props = result.records[0]._fields[0].properties;
             let label = props.name;
             delete props.name;
-            setName(label);
+            setName(label || props.objectid);
             setProperties(props);
             session.close();
         });
@@ -85,25 +87,25 @@ const NodeEditor = () => {
         }
 
         let session = driver.session();
-        session.run(statement, { id: id, newprop: newValue }).then(result => {
+        session.run(statement, { id: id, newprop: newValue }).then((result) => {
             let props = result.records[0]._fields[0].properties;
             let label = props.name;
             delete props.name;
-            setName(label);
+            setName(label || props.objectid);
             setProperties(props);
             session.close();
         });
     };
 
-    const deleteAttribute = attributeName => {
+    const deleteAttribute = (attributeName) => {
         let statement = `MATCH (n:${type} {objectid: $id}) REMOVE n.${attributeName} RETURN n`;
 
         let session = driver.session();
-        session.run(statement, { id: id }).then(result => {
+        session.run(statement, { id: id }).then((result) => {
             let props = result.records[0]._fields[0].properties;
             let label = props.name;
             delete props.name;
-            setName(label);
+            setName(label || props.objectid);
             setProperties(props);
             session.close();
         });
@@ -114,97 +116,96 @@ const NodeEditor = () => {
     }, []);
 
     return (
-        <Draggable handle={'.panel-heading'}>
-            <PoseContainer
-                visible={visible}
-                className={clsx(
-                    styles.container,
-                    context.darkMode ? styles.dark : null
-                )}
-            >
-                <Panel>
-                    <Panel.Heading>
-                        {name}
-                        <Button
-                            onClick={() => setVisible(false)}
-                            className='close'
-                            aria-label='Close'
-                        >
-                            <span aria-hidden='true'>&times;</span>
-                        </Button>
-                    </Panel.Heading>
+        <PoseContainer
+            visible={visible}
+            className={clsx(
+                styles.container,
+                context.darkMode ? styles.dark : null
+            )}
+            dragHandle={dragControl}
+        >
+            <Panel>
+                <Panel.Heading
+                    onMouseDown={(e) => {
+                        dragControl.start(e);
+                    }}
+                >
+                    {name}
+                    <Button
+                        onClick={() => setVisible(false)}
+                        className='close'
+                        aria-label='Close'
+                    >
+                        <span aria-hidden='true'>&times;</span>
+                    </Button>
+                </Panel.Heading>
 
-                    <Panel.Body>
-                        <div className='nodeEditTableContainer'>
-                            <Table>
-                                <thead align='center'>
-                                    <tr>
-                                        <td>Delete</td>
-                                        <td>Edit</td>
-                                        <td>Name</td>
-                                        <td>Value</td>
-                                    </tr>
-                                </thead>
-                                <tbody>
-                                    {Object.keys(properties).map(function(key) {
-                                        let val = properties[key];
-                                        return (
-                                            <NodeEditorRow
-                                                key={key}
-                                                attributeName={key}
-                                                val={val}
-                                                deleteHandler={deleteAttribute}
-                                                updateHandler={updateAttribute}
-                                            />
-                                        );
-                                    })}
-                                </tbody>
-                            </Table>
-                        </div>
+                <Panel.Body>
+                    <div className='nodeEditTableContainer'>
+                        <Table>
+                            <thead align='center'>
+                                <tr>
+                                    <td>Delete</td>
+                                    <td>Edit</td>
+                                    <td>Name</td>
+                                    <td>Value</td>
+                                </tr>
+                            </thead>
+                            <tbody>
+                                {Object.keys(properties).map(function (key) {
+                                    let val = properties[key];
+                                    return (
+                                        <NodeEditorRow
+                                            key={key}
+                                            attributeName={key}
+                                            val={val}
+                                            deleteHandler={deleteAttribute}
+                                            updateHandler={updateAttribute}
+                                        />
+                                    );
+                                })}
+                            </tbody>
+                        </Table>
+                    </div>
 
-                        <form
-                            onSubmit={x => x.preventDefault()}
-                            className='form-inline pull-right'
+                    <form
+                        onSubmit={(x) => x.preventDefault()}
+                        className='form-inline pull-right'
+                    >
+                        <div
+                            onFocus={() => setHasError(false)}
+                            className='form-group'
                         >
-                            <div
-                                onFocus={() => setHasError(false)}
-                                className='form-group'
+                            <input
+                                type='text'
+                                className={`${
+                                    hasError ? styles.error : ''
+                                } form-control form-override`}
+                                value={newAttrName}
+                                onChange={(e) => setNewAttrName(e.target.value)}
+                                placeholder='Internal Name'
+                                required
+                            />
+                            <select
+                                className='form-control'
+                                onChange={(e) => setNewAttrType(e.target.value)}
                             >
-                                <input
-                                    type='text'
-                                    className={`${
-                                        hasError ? styles.error : ''
-                                    } form-control form-override`}
-                                    value={newAttrName}
-                                    onChange={e =>
-                                        setNewAttrName(e.target.value)
-                                    }
-                                    placeholder='Internal Name'
-                                    required
-                                />
-                                <select
-                                    className='form-control'
-                                    onChange={e =>
-                                        setNewAttrType(e.target.value)
-                                    }
-                                >
-                                    <option value='boolean'>boolean</option>
-                                    <option value='string'>string</option>
-                                    <option value='number'>number</option>
-                                    <option value='array'>array</option>
-                                </select>
-                                <button
-                                    className='form-control formButtonFix'
-                                    onClick={() => addAttribute()}
-                                >
-                                    <span className='fa fa-plus' /> Add
-                                </button>
-                            </div>
-                        </form>
-                    </Panel.Body>
-                </Panel>
-            </PoseContainer>
-        </Draggable>
+                                <option value='boolean'>boolean</option>
+                                <option value='string'>string</option>
+                                <option value='number'>number</option>
+                                <option value='array'>array</option>
+                            </select>
+                            <button
+                                className='form-control formButtonFix'
+                                onClick={() => addAttribute()}
+                            >
+                                <span className='fa fa-plus' /> Add
+                            </button>
+                        </div>
+                    </form>
+                </Panel.Body>
+            </Panel>
+        </PoseContainer>
     );
 };
 NodeEditor.propTypes = {};
