@@ -8,31 +8,35 @@ import References from './References';
 
 const HasSession = ({ sourceName, sourceType, targetName, targetType }) => {
     return (
-        <Tabs defaultActiveKey={1} id='help-tab-container' justified>
+        <Tabs defaultActiveKey={1} id='tab-style' bsStyle='pills' justified>
             <Tab
                 eventKey={1}
-                title='Info'
+                title='GENERAL'
                 dangerouslySetInnerHTML={General(
                     sourceName,
                     sourceType,
                     targetName,
                     targetType
                 )}
+                className='helptab'
             />
             <Tab
                 eventKey={2}
-                title='Abuse Info'
+                title='ABUSE'
                 dangerouslySetInnerHTML={Abuse()}
+                className='helptab'
             />
             <Tab
                 eventKey={3}
-                title='Opsec Considerations'
+                title='OPSEC'
                 dangerouslySetInnerHTML={Opsec()}
+                className='helptab'
             />
             <Tab
                 eventKey={4}
-                title='References'
+                title='REFERENCES'
                 dangerouslySetInnerHTML={References()}
+                className='helptab'
             />
         </Tabs>
     );
