@@ -8,10 +8,10 @@ import References from './References';
 
 const MemberOf = ({ sourceName, sourceType, targetName, targetType }) => {
     return (
-        <Tabs defaultActiveKey={1} id='help-tab-container' justified>
+        <Tabs defaultActiveKey={1} id='tab-style' bsStyle='pills' justified>
             <Tab
                 eventKey={1}
-                title='Info'
+                title='GENERAL'
                 dangerouslySetInnerHTML={General(
                     sourceName,
                     sourceType,
@@ -19,19 +19,11 @@ const MemberOf = ({ sourceName, sourceType, targetName, targetType }) => {
                     targetType
                 )}
             />
-            <Tab
-                eventKey={2}
-                title='Abuse Info'
-                dangerouslySetInnerHTML={Abuse()}
-            />
-            <Tab
-                eventKey={3}
-                title='Opsec Considerations'
-                dangerouslySetInnerHTML={Opsec()}
-            />
+            <Tab eventKey={2} title='ABUSE' dangerouslySetInnerHTML={Abuse()} />
+            <Tab eventKey={3} title='OPSEC' dangerouslySetInnerHTML={Opsec()} />
             <Tab
                 eventKey={4}
-                title='References'
+                title='REFERENCES'
                 dangerouslySetInnerHTML={References()}
             />
         </Tabs>
