@@ -1,9 +1,14 @@
 import React from 'react';
-import {Highlighter} from 'react-bootstrap-typeahead';
+import { Highlighter } from 'react-bootstrap-typeahead';
+import styles from './SearchRow.module.css';
+import clsx from 'clsx';
+import { buildAzureApplicationAdmins } from '../../js/newingestion';
+import { useContext } from 'react';
+import { AppContext } from '../../AppContext';
 
-const SearchRow = (item, props) => {
+const SearchRow = ({ item, search }) => {
+    const context = useContext(AppContext);
     let searched;
-    let search = props.text;
     if (search.includes(':')) {
         searched = search.split(':')[1];
     } else {
@@ -11,7 +16,6 @@ const SearchRow = (item, props) => {
     }
     let type = item.type;
     let icon = {};
-    icon.style = { marginLeft: '10px' };
 
     switch (type) {
         case 'Group':
@@ -32,7 +36,40 @@ const SearchRow = (item, props) => {
         case 'OU':
             icon.className = 'fa fa-sitemap';
             break;
+        case 'AZUser':
+            icon.className = 'fa fa-user';
+            break;
+        case 'AZGroup':
+            icon.className = 'fa fa-users';
+            break;
+        case 'AZTenant':
+            icon.className = 'fa fa-cloud';
+            break;
+        case 'AZSubscription':
+            icon.className = 'fa fa-key';
+            break;
+        case 'AZResourceGroup':
+            icon.className = 'fa fa-cube';
+            break;
+        case 'AZVM':
+            icon.className = 'fa fa-desktop';
+            break;
+        case 'AZDevice':
+            icon.className = 'fa fa-desktop';
+            break;
+        case 'AZKeyVault':
+            icon.className = 'fa fa-lock';
+            break;
+        case 'AZApp':
+            icon.className = 'fa fa-window-restore';
+            break;
+        case 'AZServicePrincipal':
+            icon.className = 'fa fa-robot';
+            break;
     }
+
+    icon.style = { color: appStore.highResPalette.iconScheme[type].color };
+    icon.className = clsx(icon.className, styles.spacing);
 
     let name = item.name || item.objectid;
 
