@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import NodeALink from './NodeALink';
+import './NodeCypherLink.module.css';
 
 const NodeCypherLink = ({
     property,
@@ -38,11 +39,11 @@ const NodeCypherLink = ({
             objectid: target,
             domain: domain,
         })
-            .then(result => {
+            .then((result) => {
                 setValue(result.records[0].get('count'));
                 setReady(true);
             })
-            .catch(error => {
+            .catch((error) => {
                 if (
                     !error.message.includes(
                         'The transaction has been terminated'
@@ -56,24 +57,21 @@ const NodeCypherLink = ({
     }, [target]);
 
     return (
-        <>
-            <dt>{property}</dt>
-            <dd>
-                <NodeALink
-                    ready={ready}
-                    value={value}
-                    click={() => {
-                        emitter.emit(
-                            'query',
-                            `${baseQuery} RETURN p`,
-                            { objectid: target, domain: domain },
-                            start,
-                            end
-                        );
-                    }}
-                />
-            </dd>
-        </>
+        <tr
+            style={{ cursor: 'pointer' }}
+            onClick={() => {
+                emitter.emit(
+                    'query',
+                    `${baseQuery} RETURN p`,
+                    { objectid: target, domain: domain },
+                    start,
+                    end
+                );
+            }}
+        >
+            <td align='left'>{property}</td>
+            <td align='right'>{value}</td>
+        </tr>
     );
 };
 
