@@ -11,6 +11,10 @@ import NodePlayCypherLink from './Components/NodePlayCypherLink';
 import Notes from './Components/Notes';
 import { withAlert } from 'react-alert';
 import NodeGallery from './Components/NodeGallery';
+import { Table } from 'react-bootstrap';
+import styles from './NodeData.module.css';
+import { useContext } from 'react';
+import { AppContext } from '../../../AppContext';
 
 const UserNodeData = () => {
     const [visible, setVisible] = useState(false);
@@ -18,6 +22,7 @@ const UserNodeData = () => {
     const [label, setLabel] = useState(null);
     const [domain, setDomain] = useState(null);
     const [nodeProps, setNodeProps] = useState({});
+    const context = useContext(AppContext);
 
     useEffect(() => {
         emitter.on('nodeClicked', nodeClickEvent);
@@ -37,7 +42,7 @@ const UserNodeData = () => {
                 .run(`MATCH (n:User {objectid: $objectid}) RETURN n AS node`, {
                     objectid: id,
                 })
-                .then(r => {
+                .then((r) => {
                     let props = r.records[0].get('node').properties;
                     setNodeProps(props);
                     setLabel(props.name || objectid);
@@ -74,260 +79,310 @@ const UserNodeData = () => {
     return objectId === null ? (
         <div></div>
     ) : (
-        <div className={clsx(!visible && 'displaynone')}>
-            <dl className={'dl-horizontal'}>
-                <h4>{label || objectId}</h4>
-                <NodeCypherLink
-                    property='Sessions'
-                    target={objectId}
-                    baseQuery={
-                        'MATCH p=(m:Computer)-[r:HasSession]->(n:User {objectid: $objectid})'
-                    }
-                    end={label}
-                />
+        <div
+            className={clsx(
+                !visible && 'displaynone',
+                context.darkMode ? styles.dark : styles.light
+            )}
+        >
+            <div className={clsx(styles.dl)}>
+                <h5>{label || objectId}</h5>
 
-                <NodeCypherLinkComplex
-                    property='Sibling Objects in the Same OU'
-                    target={objectId}
-                    countQuery={
-                        'MATCH (o1)-[r1:Contains]->(o2:User {objectid: $objectid}) WITH o1 OPTIONAL MATCH p1=(d)-[r2:Contains*1..]->(o1) OPTIONAL MATCH p2=(o1)-[r3:Contains]->(n) WHERE n:User OR n:Computer RETURN count(distinct(n))'
-                    }
-                    graphQuery={
-                        'MATCH (o1)-[r1:Contains]->(o2:User {objectid: $objectid}) WITH o1 OPTIONAL MATCH p1=(d)-[r2:Contains*1..]->(o1) OPTIONAL MATCH p2=(o1)-[r3:Contains]->(n) WHERE n:User OR n:Computer RETURN p1,p2'
-                    }
-                />
+                <CollapsibleSection header='OVERVIEW'>
+                    <div className={styles.itemlist}>
+                        <Table>
+                            <thead></thead>
+                            <tbody className='searchable'>
+                                <NodeCypherLink
+                                    property='Sessions'
+                                    target={objectId}
+                                    baseQuery={
+                                        'MATCH p=(m:Computer)-[r:HasSession]->(n:User {objectid: $objectid})'
+                                    }
+                                    end={label}
+                                />
+                                <NodeCypherLinkComplex
+                                    property='Sibling Objects in the Same OU'
+                                    target={objectId}
+                                    countQuery={
+                                        'MATCH (o1)-[r1:Contains]->(o2:User {objectid: $objectid}) WITH o1 OPTIONAL MATCH p1=(d)-[r2:Contains*1..]->(o1) OPTIONAL MATCH p2=(o1)-[r3:Contains]->(n) WHERE n:User OR n:Computer RETURN count(distinct(n))'
+                                    }
+                                    graphQuery={
+                                        'MATCH (o1)-[r1:Contains]->(o2:User {objectid: $objectid}) WITH o1 OPTIONAL MATCH p1=(d)-[r2:Contains*1..]->(o1) OPTIONAL MATCH p2=(o1)-[r3:Contains]->(n) WHERE n:User OR n:Computer RETURN p1,p2'
+                                    }
+                                />
+                                <NodeCypherLink
+                                    property='Reachable High Value Targets'
+                                    target={objectId}
+                                    baseQuery={
+                                        'MATCH (m:User {objectid: $objectid}),(n {highvalue:true}),p=shortestPath((m)-[r*1..]->(n)) WHERE NONE (r IN relationships(p) WHERE type(r)= "GetChanges") AND NONE (r in relationships(p) WHERE type(r)="GetChangesAll") AND NOT m=n'
+                                    }
+                                    start={label}
+                                />
+                                <NodeCypherLinkComplex
+                                    property='Effective Inbound GPOs'
+                                    target={objectId}
+                                    countQuery={
+                                        'MATCH (c:User {objectid: $objectid}) OPTIONAL MATCH p1 = (g1:GPO)-[r1:GpLink {enforced:true}]->(container1)-[r2:Contains*1..]->(c) OPTIONAL MATCH p2 = (g2:GPO)-[r3:GpLink {enforced:false}]->(container2)-[r4:Contains*1..]->(c) WHERE NONE (x in NODES(p2) WHERE x.blocksinheritance = true AND x:OU AND NOT (g2)-->(x)) WITH COLLECT(g1) + COLLECT(g2) AS tempVar UNWIND tempVar AS GPOs RETURN COUNT(DISTINCT(GPOs))'
+                                    }
+                                    graphQuery={
+                                        'MATCH (c:User {objectid: $objectid}) OPTIONAL MATCH p1 = (g1:GPO)-[r1:GpLink {enforced:true}]->(container1)-[r2:Contains*1..]->(c) OPTIONAL MATCH p2 = (g2:GPO)-[r3:GpLink {enforced:false}]->(container2)-[r4:Contains*1..]->(c) WHERE NONE (x in NODES(p2) WHERE x.blocksinheritance = true AND x:OU AND NOT (g2)-->(x)) RETURN p1,p2'
+                                    }
+                                />
+                                <NodeCypherNoNumberLink
+                                    target={objectId}
+                                    property='See user within Domain/OU Tree'
+                                    query='MATCH p = (d:Domain)-[r:Contains*1..]->(u:User {objectid: $objectid}) RETURN p'
+                                />
+                            </tbody>
+                        </Table>
+                    </div>
+                </CollapsibleSection>
 
-                <NodeCypherLink
-                    property='Reachable High Value Targets'
-                    target={objectId}
-                    baseQuery={
-                        'MATCH (m:User {objectid: $objectid}),(n {highvalue:true}),p=shortestPath((m)-[r*1..]->(n)) WHERE NONE (r IN relationships(p) WHERE type(r)= "GetChanges") AND NONE (r in relationships(p) WHERE type(r)="GetChangesAll") AND NOT m=n'
-                    }
-                    start={label}
-                />
+                <hr></hr>
 
-                <NodeCypherLinkComplex
-                    property='Effective Inbound GPOs'
-                    target={objectId}
-                    countQuery={
-                        'MATCH (c:User {objectid: $objectid}) OPTIONAL MATCH p1 = (g1:GPO)-[r1:GpLink {enforced:true}]->(container1)-[r2:Contains*1..]->(c) OPTIONAL MATCH p2 = (g2:GPO)-[r3:GpLink {enforced:false}]->(container2)-[r4:Contains*1..]->(c) WHERE NONE (x in NODES(p2) WHERE x.blocksinheritance = true AND x:OU AND NOT (g2)-->(x)) WITH COLLECT(g1) + COLLECT(g2) AS tempVar UNWIND tempVar AS GPOs RETURN COUNT(DISTINCT(GPOs))'
-                    }
-                    graphQuery={
-                        'MATCH (c:User {objectid: $objectid}) OPTIONAL MATCH p1 = (g1:GPO)-[r1:GpLink {enforced:true}]->(container1)-[r2:Contains*1..]->(c) OPTIONAL MATCH p2 = (g2:GPO)-[r3:GpLink {enforced:false}]->(container2)-[r4:Contains*1..]->(c) WHERE NONE (x in NODES(p2) WHERE x.blocksinheritance = true AND x:OU AND NOT (g2)-->(x)) RETURN p1,p2'
-                    }
-                />
-
-                <NodeCypherNoNumberLink
-                    target={objectId}
-                    property='See user within Domain/OU Tree'
-                    query='MATCH p = (d:Domain)-[r:Contains*1..]->(u:User {objectid: $objectid}) RETURN p'
-                />
                 <MappedNodeProps
                     displayMap={displayMap}
                     properties={nodeProps}
                     label={label}
                 />
+
+                <hr></hr>
+
                 <ExtraNodeProps
                     displayMap={displayMap}
                     properties={nodeProps}
                     label={label}
                 />
 
-                <CollapsibleSection header={'Group Membership'}>
-                    <NodeCypherLink
-                        property='First Degree Group Memberships'
-                        target={objectId}
-                        baseQuery={
-                            'MATCH (m:User {objectid: $objectid}), (n:Group), p=(m)-[:MemberOf]->(n)'
-                        }
-                        start={label}
-                    />
+                <hr></hr>
 
-                    <NodeCypherLink
-                        property='Unrolled Group Membership'
-                        target={objectId}
-                        baseQuery={
-                            'MATCH p = (m:User {objectid: $objectid})-[r:MemberOf*1..]->(n:Group)'
-                        }
-                        start={label}
-                        distinct
-                    />
-
-                    <NodeCypherLink
-                        property='Foreign Group Membership'
-                        target={objectId}
-                        baseQuery={
-                            'MATCH (m:User {objectid: $objectid}) MATCH (n:Group) WHERE NOT m.domain=n.domain MATCH p=(m)-[r:MemberOf*1..]->(n)'
-                        }
-                        start={label}
-                        domain={domain}
-                    />
+                <CollapsibleSection header={'GROUP MEMBERSHIP'}>
+                    <div className={styles.itemlist}>
+                        <Table>
+                            <thead></thead>
+                            <tbody className='searchable'>
+                                <NodeCypherLink
+                                    property='First Degree Group Memberships'
+                                    target={objectId}
+                                    baseQuery={
+                                        'MATCH (m:User {objectid: $objectid}), (n:Group), p=(m)-[:MemberOf]->(n)'
+                                    }
+                                    start={label}
+                                />
+                                <NodeCypherLink
+                                    property='Unrolled Group Membership'
+                                    target={objectId}
+                                    baseQuery={
+                                        'MATCH p = (m:User {objectid: $objectid})-[r:MemberOf*1..]->(n:Group)'
+                                    }
+                                    start={label}
+                                    distinct
+                                />
+                                <NodeCypherLink
+                                    property='Foreign Group Membership'
+                                    target={objectId}
+                                    baseQuery={
+                                        'MATCH (m:User {objectid: $objectid}) MATCH (n:Group) WHERE NOT m.domain=n.domain MATCH p=(m)-[r:MemberOf*1..]->(n)'
+                                    }
+                                    start={label}
+                                    domain={domain}
+                                />
+                            </tbody>
+                        </Table>
+                    </div>
                 </CollapsibleSection>
 
-                <CollapsibleSection header={'Local Admin Rights'}>
-                    <NodeCypherLink
-                        property='First Degree Local Admin'
-                        target={objectId}
-                        baseQuery={
-                            'MATCH p=(m:User {objectid: $objectid})-[r:AdminTo]->(n:Computer)'
-                        }
-                        start={label}
-                        distinct
-                    />
+                <hr></hr>
 
-                    <NodeCypherLink
-                        property='Group Delegated Local Admin Rights'
-                        target={objectId}
-                        baseQuery={
-                            'MATCH p=(m:User {objectid: $objectid})-[r1:MemberOf*1..]->(g:Group)-[r2:AdminTo]->(n:Computer)'
-                        }
-                        start={label}
-                        distinct
-                    />
-
-                    <NodePlayCypherLink
-                        property='Derivative Local Admin Rights'
-                        target={objectId}
-                        baseQuery={
-                            'MATCH p=shortestPath((m:User {objectid: $objectid})-[r:HasSession|AdminTo|MemberOf*1..]->(n:Computer))'
-                        }
-                        start={label}
-                        distinct
-                    />
+                <CollapsibleSection header={'LOCAL ADMIN RIGHTS'}>
+                    <div className={styles.itemlist}>
+                        <Table>
+                            <thead></thead>
+                            <tbody className='searchable'>
+                                <NodeCypherLink
+                                    property='First Degree Local Admin'
+                                    target={objectId}
+                                    baseQuery={
+                                        'MATCH p=(m:User {objectid: $objectid})-[r:AdminTo]->(n:Computer)'
+                                    }
+                                    start={label}
+                                    distinct
+                                />
+                                <NodeCypherLink
+                                    property='Group Delegated Local Admin Rights'
+                                    target={objectId}
+                                    baseQuery={
+                                        'MATCH p=(m:User {objectid: $objectid})-[r1:MemberOf*1..]->(g:Group)-[r2:AdminTo]->(n:Computer)'
+                                    }
+                                    start={label}
+                                    distinct
+                                />
+                                <NodePlayCypherLink
+                                    property='Derivative Local Admin Rights'
+                                    target={objectId}
+                                    baseQuery={
+                                        'MATCH p=shortestPath((m:User {objectid: $objectid})-[r:HasSession|AdminTo|MemberOf*1..]->(n:Computer))'
+                                    }
+                                    start={label}
+                                    distinct
+                                />
+                            </tbody>
+                        </Table>
+                    </div>
                 </CollapsibleSection>
 
-                <CollapsibleSection header={'Execution Privileges'}>
-                    <NodeCypherLink
-                        property='First Degree RDP Privileges'
-                        target={objectId}
-                        baseQuery={
-                            'MATCH p=(m:User {objectid: $objectid})-[r:CanRDP]->(n:Computer)'
-                        }
-                        start={label}
-                        distinct
-                    />
+                <hr></hr>
 
-                    <NodeCypherLink
-                        property='Group Delegated RDP Privileges'
-                        target={objectId}
-                        baseQuery={
-                            'MATCH p=(m:User {objectid: $objectid})-[r1:MemberOf*1..]->(g:Group)-[r2:CanRDP]->(n:Computer)'
-                        }
-                        start={label}
-                        distinct
-                    />
-
-                    <NodeCypherLink
-                        property='First Degree DCOM Privileges'
-                        target={objectId}
-                        baseQuery={
-                            'MATCH p=(m:User {objectid: $objectid})-[r:ExecuteDCOM]->(n:Computer)'
-                        }
-                        start={label}
-                        distinct
-                    />
-
-                    <NodeCypherLink
-                        property='Group Delegated DCOM Privileges'
-                        target={objectId}
-                        baseQuery={
-                            'MATCH p=(m:User {objectid: $objectid})-[r1:MemberOf*1..]->(g:Group)-[r2:ExecuteDCOM]->(n:Computer)'
-                        }
-                        start={label}
-                        distinct
-                    />
-
-                    <NodeCypherLink
-                        property='SQL Admin Rights'
-                        target={objectId}
-                        baseQuery={
-                            'MATCH p=(m:User {objectid: $objectid})-[r:SQLAdmin]->(n:Computer)'
-                        }
-                        start={label}
-                        distinct
-                    />
-
-                    <NodeCypherLink
-                        property='Constrained Delegation Privileges'
-                        target={objectId}
-                        baseQuery={
-                            'MATCH p=(m:User {objectid: $objectid})-[r:AllowedToDelegate]->(n:Computer)'
-                        }
-                        start={label}
-                        distinct
-                    />
+                <CollapsibleSection header={'EXECUTION RIGHTS'}>
+                    <div className={styles.itemlist}>
+                        <Table>
+                            <thead></thead>
+                            <tbody className='searchable'>
+                                <NodeCypherLink
+                                    property='First Degree RDP Privileges'
+                                    target={objectId}
+                                    baseQuery={
+                                        'MATCH p=(m:User {objectid: $objectid})-[r:CanRDP]->(n:Computer)'
+                                    }
+                                    start={label}
+                                    distinct
+                                />
+                                <NodeCypherLink
+                                    property='Group Delegated RDP Privileges'
+                                    target={objectId}
+                                    baseQuery={
+                                        'MATCH p=(m:User {objectid: $objectid})-[r1:MemberOf*1..]->(g:Group)-[r2:CanRDP]->(n:Computer)'
+                                    }
+                                    start={label}
+                                    distinct
+                                />
+                                <NodeCypherLink
+                                    property='First Degree DCOM Privileges'
+                                    target={objectId}
+                                    baseQuery={
+                                        'MATCH p=(m:User {objectid: $objectid})-[r:ExecuteDCOM]->(n:Computer)'
+                                    }
+                                    start={label}
+                                    distinct
+                                />
+                                <NodeCypherLink
+                                    property='Group Delegated DCOM Privileges'
+                                    target={objectId}
+                                    baseQuery={
+                                        'MATCH p=(m:User {objectid: $objectid})-[r1:MemberOf*1..]->(g:Group)-[r2:ExecuteDCOM]->(n:Computer)'
+                                    }
+                                    start={label}
+                                    distinct
+                                />
+                                <NodeCypherLink
+                                    property='SQL Admin Rights'
+                                    target={objectId}
+                                    baseQuery={
+                                        'MATCH p=(m:User {objectid: $objectid})-[r:SQLAdmin]->(n:Computer)'
+                                    }
+                                    start={label}
+                                    distinct
+                                />
+                                <NodeCypherLink
+                                    property='Constrained Delegation Privileges'
+                                    target={objectId}
+                                    baseQuery={
+                                        'MATCH p=(m:User {objectid: $objectid})-[r:AllowedToDelegate]->(n:Computer)'
+                                    }
+                                    start={label}
+                                    distinct
+                                />
+                            </tbody>
+                        </Table>
+                    </div>
                 </CollapsibleSection>
 
-                <CollapsibleSection header={'Outbound Object Control'}>
-                    <NodeCypherLink
-                        property='First Degree Object Control'
-                        target={objectId}
-                        baseQuery={
-                            'MATCH p=(u:User {objectid: $objectid})-[r1]->(n) WHERE r1.isacl=true'
-                        }
-                        end={label}
-                        distinct
-                    />
+                <hr></hr>
 
-                    <NodeCypherLink
-                        property='Group Delegated Object Control'
-                        target={objectId}
-                        baseQuery={
-                            'MATCH p=(u:User {objectid: $objectid})-[r1:MemberOf*1..]->(g:Group)-[r2]->(n) WHERE r2.isacl=true'
-                        }
-                        start={label}
-                        distinct
-                    />
-
-                    <NodePlayCypherLink
-                        property='Transitive Object Control'
-                        target={objectId}
-                        baseQuery={
-                            'MATCH (n) WHERE NOT n.objectid=$objectid MATCH p=shortestPath((u:User {objectid: $objectid})-[r1:MemberOf|AddMember|AllExtendedRights|ForceChangePassword|GenericAll|GenericWrite|WriteDacl|WriteOwner|Owns*1..]->(n))'
-                        }
-                        start={label}
-                        distinct
-                    />
+                <CollapsibleSection header={'OUTBOUND CONTROL RIGHTS'}>
+                    <div className={styles.itemlist}>
+                        <Table>
+                            <thead></thead>
+                            <tbody className='searchable'>
+                                <NodeCypherLink
+                                    property='First Degree Object Control'
+                                    target={objectId}
+                                    baseQuery={
+                                        'MATCH p=(u:User {objectid: $objectid})-[r1]->(n) WHERE r1.isacl=true'
+                                    }
+                                    end={label}
+                                    distinct
+                                />
+                                <NodeCypherLink
+                                    property='Group Delegated Object Control'
+                                    target={objectId}
+                                    baseQuery={
+                                        'MATCH p=(u:User {objectid: $objectid})-[r1:MemberOf*1..]->(g:Group)-[r2]->(n) WHERE r2.isacl=true'
+                                    }
+                                    start={label}
+                                    distinct
+                                />
+                                <NodePlayCypherLink
+                                    property='Transitive Object Control'
+                                    target={objectId}
+                                    baseQuery={
+                                        'MATCH (n) WHERE NOT n.objectid=$objectid MATCH p=shortestPath((u:User {objectid: $objectid})-[r1:MemberOf|AddMember|AllExtendedRights|ForceChangePassword|GenericAll|GenericWrite|WriteDacl|WriteOwner|Owns*1..]->(n))'
+                                    }
+                                    start={label}
+                                    distinct
+                                />
+                            </tbody>
+                        </Table>
+                    </div>
                 </CollapsibleSection>
 
-                <CollapsibleSection header={'Inbound Object Control'}>
-                    <NodeCypherLink
-                        property='Explicit Object Controllers'
-                        target={objectId}
-                        baseQuery={
-                            'MATCH p=(n)-[r]->(u1:User {objectid: $objectid}) WHERE r.isacl=true'
-                        }
-                        end={label}
-                        distinct
-                    />
+                <hr></hr>
 
-                    <NodeCypherLink
-                        property='Unrolled Object Controllers'
-                        target={objectId}
-                        baseQuery={
-                            'MATCH p=(n)-[r:MemberOf*1..]->(g:Group)-[r1:AddMember|AllExtendedRights|GenericAll|GenericWrite|WriteDacl|WriteOwner|Owns]->(u:User {objectid: $objectid}) WITH LENGTH(p) as pathLength, p, n WHERE NONE (x in NODES(p)[1..(pathLength-1)] WHERE x.objectid = u.objectid) AND NOT n.objectid = u.objectid'
-                        }
-                        end={label}
-                        distinct
-                    />
-
-                    <NodePlayCypherLink
-                        property='Transitive Object Controllers'
-                        target={objectId}
-                        baseQuery={
-                            'MATCH (n) WHERE NOT n.objectid=$objectid MATCH p = shortestPath((n)-[r1:MemberOf|AllExtendedRights|ForceChangePassword|GenericAll|GenericWrite|WriteDacl|WriteOwner*1..]->(u1:User {objectid: $objectid}))'
-                        }
-                        end={label}
-                        distinct
-                    />
+                <CollapsibleSection header={'INBOUND CONTROL RIGHTS'}>
+                    <div className={styles.itemlist}>
+                        <Table>
+                            <thead></thead>
+                            <tbody className='searchable'>
+                                <NodeCypherLink
+                                    property='Explicit Object Controllers'
+                                    target={objectId}
+                                    baseQuery={
+                                        'MATCH p=(n)-[r]->(u1:User {objectid: $objectid}) WHERE r.isacl=true'
+                                    }
+                                    end={label}
+                                    distinct
+                                />
+                                <NodeCypherLink
+                                    property='Unrolled Object Controllers'
+                                    target={objectId}
+                                    baseQuery={
+                                        'MATCH p=(n)-[r:MemberOf*1..]->(g:Group)-[r1:AddMember|AllExtendedRights|GenericAll|GenericWrite|WriteDacl|WriteOwner|Owns]->(u:User {objectid: $objectid}) WITH LENGTH(p) as pathLength, p, n WHERE NONE (x in NODES(p)[1..(pathLength-1)] WHERE x.objectid = u.objectid) AND NOT n.objectid = u.objectid'
+                                    }
+                                    end={label}
+                                    distinct
+                                />
+                                <NodePlayCypherLink
+                                    property='Transitive Object Controllers'
+                                    target={objectId}
+                                    baseQuery={
+                                        'MATCH (n) WHERE NOT n.objectid=$objectid MATCH p = shortestPath((n)-[r1:MemberOf|AllExtendedRights|ForceChangePassword|GenericAll|GenericWrite|WriteDacl|WriteOwner*1..]->(u1:User {objectid: $objectid}))'
+                                    }
+                                    end={label}
+                                    distinct
+                                />
+                            </tbody>
+                        </Table>
+                    </div>
                 </CollapsibleSection>
-                <Notes objectid={objectId} type={'User'} />
+
+                {/* <Notes objectid={objectId} type={'User'} />
                 <NodeGallery
                     objectid={objectId}
                     type={'User'}
                     visible={visible}
-                />
-            </dl>
+                /> */}
+            </div>
         </div>
     );
 };
