@@ -1,6 +1,23 @@
 import { groupBy } from 'lodash/collection';
 
-var labels = ['OU', 'GPO', 'User', 'Computer', 'Group', 'Domain'];
+var labels = [
+    'OU',
+    'GPO',
+    'User',
+    'Computer',
+    'Group',
+    'Domain',
+    'AZApp',
+    'AZDevice',
+    'AZGroup',
+    'AZKeyVault',
+    'AZResourceGroup',
+    'AZServicePrincipal',
+    'AZSubscription',
+    'AZTenant',
+    'AZUser',
+    'AZVM',
+];
 
 export function generateUniqueId(sigmaInstance, isNode) {
     var i = Math.floor(Math.random() * (100000 - 10 + 1)) + 10;
@@ -31,20 +48,15 @@ function getRealLabel(label) {
 export function buildSearchQuery(searchterm) {
     if (searchterm.includes(':')) {
         let [type, term] = searchterm.split(':');
-        term = escapeRegExp(term);
-        let t = '(?i).*' + term + '.*';
         type = getRealLabel(type);
 
-        let statement = `MATCH (n:${type}) WHERE n.name =~ $name OR n.objectid =~ $name RETURN n LIMIT 10`;
+        let statement = `MATCH (n:${type}) WHERE n.name CONTAINS $name OR n.objectid CONTAINS $name RETURN n LIMIT 10`;
 
-        return [statement, t];
+        return [statement, term.toUpperCase()];
     } else {
-        let q = escapeRegExp(searchterm);
-        let t = '(?i).*' + q + '.*';
-
         return [
-            'MATCH (n) WHERE n.name =~ $name OR n.objectid =~ $name RETURN n LIMIT 10',
-            t,
+            'MATCH (n:Base) WHERE n.name CONTAINS $name OR n.objectid CONTAINS $name RETURN n LIMIT 10',
+            searchterm.toUpperCase(),
         ];
     }
 }
@@ -73,8 +85,8 @@ export function findGraphPath(sigmaInstance, reverse, nodeid, traversed) {
             ? sigmaInstance.graph.inboundNodes(nodeid)
             : sigmaInstance.graph.outboundNodes(nodeid);
         //Loop over the nodes near us and the edges connecting to those nodes
-        $.each(nodes, function(index, node) {
-            $.each(edges, function(index, edge) {
+        $.each(nodes, function (index, node) {
+            $.each(edges, function (index, edge) {
                 var check = reverse ? edge.source : edge.target;
                 //If an edge is pointing in the right direction, set its color
                 //Push the edge into our store and then
@@ -102,7 +114,7 @@ function deleteSessions() {
         .run(
             'MATCH ()-[r:HasSession]-() WITH r LIMIT 100000 DELETE r RETURN count(r)'
         )
-        .then(function(results) {
+        .then(function (results) {
             session.close();
             emitter.emit('refreshDBData');
             var count = results.records[0]._fields[0];
@@ -123,7 +135,7 @@ function deleteEdges() {
     var session = driver.session();
     session
         .run('MATCH ()-[r]-() WITH r LIMIT 100000 DELETE r RETURN count(r)')
-        .then(function(results) {
+        .then(function (results) {
             emitter.emit('refreshDBData');
             session.close();
             var count = results.records[0]._fields[0];
@@ -139,7 +151,7 @@ function deleteNodes() {
     var session = driver.session();
     session
         .run('MATCH (n) WITH n LIMIT 100000 DELETE n RETURN count(n)')
-        .then(function(results) {
+        .then(function (results) {
             emitter.emit('refreshDBData');
             session.close();
             var count = results.records[0]._fields[0];
@@ -154,10 +166,16 @@ function deleteNodes() {
 function grabConstraints() {
     var session = driver.session();
     let constraints = [];
-    session.run('CALL db.constraints').then(function(results) {
-        $.each(results.records, function(index, container) {
+    session.run('CALL db.constraints').then(function (results) {
+        $.each(results.records, function (index, container) {
             let constraint = container._fields[0];
-            let query = 'DROP ' + constraint;
+            let query;
+            if (neoVersion.startsWith('3.')) {
+                query = 'DROP ' + constraint;
+            } else {
+                query = 'DROP CONSTRAINT ' + constraint;
+            }
+
             constraints.push(query);
         });
 
@@ -171,7 +189,7 @@ function dropConstraints(constraints) {
     if (constraints.length > 0) {
         let constraint = constraints.shift();
         let session = driver.session();
-        session.run(constraint).then(function() {
+        session.run(constraint).then(function () {
             dropConstraints(constraints);
             session.close();
         });
@@ -184,10 +202,17 @@ function grabIndexes() {
     var session = driver.session();
     let constraints = [];
 
-    session.run('CALL db.indexes').then(function(results) {
-        $.each(results.records, function(index, container) {
-            let constraint = container._fields[0];
-            let query = 'DROP ' + constraint;
+    session.run('CALL db.indexes').then(function (results) {
+        $.each(results.records, function (index, container) {
+            let query;
+            if (neoVersion.startsWith('3.')) {
+                let constraint = container._fields[0];
+                query = 'DROP ' + constraint;
+            } else {
+                let constraint = container._fields[1];
+                query = 'DROP INDEX ' + constraint;
+            }
+
             constraints.push(query);
         });
 
@@ -201,7 +226,7 @@ function dropIndexes(indexes) {
     if (indexes.length > 0) {
         let constraint = indexes.shift();
         let session = driver.session();
-        session.run(constraint).then(function() {
+        session.run(constraint).then(function () {
             dropConstraints(indexes);
             session.close();
         });
@@ -214,20 +239,20 @@ export async function addConstraints() {
     let session = driver.session();
     await session
         .run('CREATE CONSTRAINT ON (c:Base) ASSERT c.objectid IS UNIQUE')
-        .catch(_ => {});
-    await session.run('CREATE INDEX ON :User(name)').catch(_ => {});
-    await session.run('CREATE INDEX ON :User(objectid)').catch(_ => {});
-    await session.run('CREATE INDEX ON :Group(name)').catch(_ => {});
-    await session.run('CREATE INDEX ON :Group(objectid)').catch(_ => {});
-    await session.run('CREATE INDEX ON :Computer(name)').catch(_ => {});
-    await session.run('CREATE INDEX ON :Computer(objectid)').catch(_ => {});
-    await session.run('CREATE INDEX ON :GPO(name)').catch(_ => {});
-    await session.run('CREATE INDEX ON :GPO(objectid)').catch(_ => {});
-    await session.run('CREATE INDEX ON :Domain(name)').catch(_ => {});
-    await session.run('CREATE INDEX ON :Domain(objectid)').catch(_ => {});
-    await session.run('CREATE INDEX ON :OU(name)').catch(_ => {});
-    await session.run('CREATE INDEX ON :OU(objectid)').catch(_ => {});
-    await session.run('CREATE INDEX ON :Base(name)').catch(_ => {});
+        .catch((_) => {});
+    await session.run('CREATE INDEX ON :User(name)').catch((_) => {});
+    await session.run('CREATE INDEX ON :User(objectid)').catch((_) => {});
+    await session.run('CREATE INDEX ON :Group(name)').catch((_) => {});
+    await session.run('CREATE INDEX ON :Group(objectid)').catch((_) => {});
+    await session.run('CREATE INDEX ON :Computer(name)').catch((_) => {});
+    await session.run('CREATE INDEX ON :Computer(objectid)').catch((_) => {});
+    await session.run('CREATE INDEX ON :GPO(name)').catch((_) => {});
+    await session.run('CREATE INDEX ON :GPO(objectid)').catch((_) => {});
+    await session.run('CREATE INDEX ON :Domain(name)').catch((_) => {});
+    await session.run('CREATE INDEX ON :Domain(objectid)').catch((_) => {});
+    await session.run('CREATE INDEX ON :OU(name)').catch((_) => {});
+    await session.run('CREATE INDEX ON :OU(objectid)').catch((_) => {});
+    await session.run('CREATE INDEX ON :Base(name)').catch((_) => {});
     session.close();
 
     emitter.emit('hideDBClearModal');
